Refactor DBUtils
