+++ dpc-attribution/pom.xml
@@ -122,12 +122,6 @@
             <artifactId>mockito-junit-jupiter</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>gov.cms.dpc</groupId>
-            <artifactId>dpc-testing</artifactId>
-            <version>0.4.0-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
     </dependencies>
 
     <build>
