+++ dpc-common/src/main/java/gov/cms/dpc/common/exceptions/NotATestEnvironmentException.java
@@ -1,4 +1,4 @@
-package gov.cms.dpc.testing.exceptions;
+package gov.cms.dpc.common.exceptions;
 
 /**
  * Thrown when you're running a destructive test in a persisted environment
