+++ dpc-common/src/test/java/gov/cms/dpc/common/utils/DBUtilsUnitTest.java
@@ -1,6 +1,6 @@
-package gov.cms.dpc.testing.utils;
+package gov.cms.dpc.common.utils;
 
-import gov.cms.dpc.testing.exceptions.NotATestEnvironmentException;
+import gov.cms.dpc.common.exceptions.NotATestEnvironmentException;
 import org.jooq.DSLContext;
 import org.jooq.SQLDialect;
 import org.jooq.impl.DefaultDSLContext;
