+++ dpc-consent/pom.xml
@@ -137,12 +137,6 @@
             <version>1.6.13</version>
             <scope>compile</scope>
         </dependency>
-        <dependency>
-            <groupId>gov.cms.dpc</groupId>
-            <artifactId>dpc-testing</artifactId>
-            <version>0.4.0-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
     </dependencies>
 
     <build>
